requirements.txt
# Зависимости для работы со справочником Контрагенты из 1С
# Python 3.x требуется

# Для работы с COM и подключения к 1С:Предприятие
pywin32>=306

# Для работы с .env файлами
python-dotenv>=1.0.0

# SQLite встроен в Python, дополнительные библиотеки не требуются

# Для веб-интерфейса
Flask>=2.3.0
flask-cors>=4.0.0

# Для отправки уведомлений через Telegram
requests>=2.31.0

